+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [4.0.0.0] - 2026-09-17
+
 ### Fixed
 
 - A `{-# LINE n "file" #-}` pragma now applies its file name. The file name was